+++ src/main/java/intellimate/izouSDK/contentgenerator/ContentGeneratorImpl.java
@@ -3,6 +3,7 @@ package intellimate.izouSDK.contentgenerator;
 import intellimate.izou.events.Event;
 import intellimate.izou.resource.Resource;
 import intellimate.izou.system.Context;
+import intellimate.izou.system.context.ContextImplementation;
 import intellimate.izou.identification.IdentificationManager;
 
 import java.util.List;
@@ -18,7 +19,7 @@ import java.util.Optional;
 public abstract class ContentGeneratorImpl implements ContentGenerator {
     //stores the ID of the ContentGenerator
     private final String contentGeneratorID;
-    private final Context context;
+    private final ContextImplementation context;
     private IdentificationManager identificationManager = IdentificationManager.getInstance();
 
     /**
@@ -35,7 +36,7 @@ public abstract class ContentGeneratorImpl implements ContentGenerator {
      * @param eventIDName the event id name
      * @param eventID the actual event id
      */
-    public ContentGeneratorImpl(String id, Context context, String eventDescription, String eventIDName,
+    public ContentGeneratorImpl(String id, ContextImplementation context, String eventDescription, String eventIDName,
                                 String eventID) {
         this.contentGeneratorID = id;
         this.context = context;
@@ -91,7 +92,7 @@ public abstract class ContentGeneratorImpl implements ContentGenerator {
      *
      * @return an instance of Context.
      */
-    public intellimate.izou.system.context.Context getContext() {
+    public Context getContext() {
         return context;
     }
