+++ src/test/java/intellimate/izou/fullplugintesting/TestCG.java
@@ -2,7 +2,7 @@ package intellimate.izou.fullplugintesting;
 
 import intellimate.izouSDK.resource.ResourceImpl;
 import intellimate.izouSDK.contentgenerator.ContentGeneratorImpl;
-import intellimate.izou.system.Context;
+import intellimate.izou.system.context.ContextImplementation;
 import intellimate.izou.events.Event;
 import intellimate.izou.resource.Resource;
 
@@ -13,7 +13,7 @@ import java.util.Optional;
 @SuppressWarnings("SameParameterValue")
 public class TestCG extends ContentGeneratorImpl {
 
-    public TestCG(Context context) {
+    public TestCG(ContextImplementation context) {
         super(TestCG.class.getCanonicalName(), context);
     }
