+++ src/test/java/intellimate/izou/fullplugintesting/TestOE.java
@@ -3,6 +3,7 @@ package intellimate.izou.fullplugintesting;
 import intellimate.izou.events.Event;
 import intellimate.izou.output.OutputExtensionimpl;
 import intellimate.izou.resource.Resource;
+import intellimate.izou.system.Context;
 
 import java.util.stream.Collectors;
 
@@ -12,7 +13,7 @@ import java.util.stream.Collectors;
 @SuppressWarnings("SameParameterValue")
 public class TestOE extends OutputExtensionimpl<TestOD> {
 
-    public TestOE(String id, intellimate.izou.system.context.Context context) {
+    public TestOE(String id, Context context) {
         super(id, context);
         setPluginId("test-OP");
         addResourceIdToWishList("test_ID");
