+++ src/test/java/intellimate/izou/testHelper/IzouTest.java
@@ -3,6 +3,7 @@ package intellimate.izou.testHelper;
 import intellimate.izou.addon.AddOn;
 import intellimate.izou.events.Event;
 import intellimate.izou.system.Context;
+import intellimate.izou.system.context.ContextImplementation;
 import intellimate.izou.system.javafx.JavaFXInitializer;
 import intellimate.izouSDK.events.EventImpl;
 import intellimate.izou.events.MultipleEventsException;
@@ -35,7 +36,7 @@ public class IzouTest implements Identifiable {
     private static final class Lock { }
     private final Object lock = new Lock();
     private TestAddOn testAddOn = new TestAddOn(getID());
-    private intellimate.izou.system.context.Context context;
+    private Context context;
 
     /**
      * creates a new instance of IzouTest
@@ -50,7 +51,7 @@ public class IzouTest implements Identifiable {
             }
             this.main = staticMain;
         }
-        context = new Context(testAddOn, main, "debug");
+        context = new ContextImplementation(testAddOn, main, "debug");
         this.id = id;
         identificationManager = IdentificationManager.getInstance();
         identificationManager.registerIdentification(this);
@@ -172,7 +173,7 @@ public class IzouTest implements Identifiable {
         }
     }
 
-    public intellimate.izou.system.context.Context getContext() {
+    public Context getContext() {
         return context;
     }
